requirements.txt
djangorestframework==3.14.0
pip-chill==1.0.3
djangorestframework==3.14.0
pillow==10.0.0
pip-chill==1.0.3
